Propagate ALTER ROLE .. SET statements
In PostgreSQL, user defaults for config parameters can be changed by
ALTER ROLE .. SET statements. We wish to propagate those defaults
accross the Citus cluster so that the behaviour will be similar in
different workers.

The only unsupported SET statement is SET FROM CURRENT, and the user
needs to provide a concrete value to make sure that the value is the same
accross the whole cluster

The defaults can either be set in a specific database, or the whole cluster.